Rename bootstring To Core
Move bootstring code into the `org.typelevel.idna4s.core.bootstring` package.

There are several utility functions that both the IDNA core code and `bootstring` will want to access and they will share all dependencies, so I don't think we need to keep them as separate modules.
