--- bootstring/src/main/scala-2/cats/bootstring/syntax/BiasSyntax.scala
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2022 Typelevel
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package org.typelevel.idna4s.bootstring.syntax
-
-import org.typelevel.idna4s.bootstring._
-import org.typelevel.literally.Literally
-
-private[syntax] trait BiasSyntax {
-  implicit class BiasContext(val sc: StringContext) {
-    def bias(args: Any*): Bias = macro BiasSyntax.bias.make
-  }
-}
-
-private object BiasSyntax {
-
-  private object bias extends Literally[Bias] {
-    def validate(c: Context)(s: String): Either[String, c.Expr[Bias]] = {
-      import c.universe._
-
-      Bias.fromString(s).map(_ => c.Expr(q"Bias.unsafeFromString($s)"))
-    }
-
-    def make(c: Context)(args: c.Expr[Any]*): c.Expr[Bias] =
-      apply(c)(args: _*)
-  }
-}
