--- bootstring/src/main/scala-3/cats/bootstring/syntax/BiasSyntax.scala
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2022 Typelevel
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package org.typelevel.idna4s.bootstring.syntax
-
-import org.typelevel.idna4s.bootstring._
-import scala.language.future
-import scala.quoted.*
-
-private[syntax] trait BiasSyntax {
-  extension (inline ctx: StringContext) {
-    inline def bias(inline args: Any*): Bias =
-      BiasSyntax.biasLiteral(ctx, args)
-  }
-}
-
-private object BiasSyntax {
-
-  private def biasLiteralExpr(sc: Expr[StringContext], args: Expr[Seq[Any]])(
-      using q: Quotes): Expr[Bias] =
-    sc.value match {
-      case Some(sc) if sc.parts.size == 1 =>
-        val value: String = sc.parts.head
-        Bias
-          .fromString(value)
-          .fold(
-            e => {
-              quotes.reflect.report.throwError(e)
-            },
-            _ => '{ Bias.unsafeFromString(${ Expr(value) }) }
-          )
-      case Some(_) =>
-        quotes.reflect.report.throwError("StringContext must be a single string literal")
-      case None =>
-        quotes.reflect.report.throwError("StringContext args must be statically known")
-    }
-
-  inline def biasLiteral(inline sc: StringContext, inline args: Any*): Bias =
-    ${ biasLiteralExpr('sc, 'args) }
-}
